Divaldil

+++ src/modules/orcamentos/orcamentos.service.ts
@@ -6,7 +6,10 @@ import { ProdutosService } from '../produtos/produtos.service';
 
 @Injectable()
 export class OrcamentosService {
-  constructor(private readonly prismaService: PrismaService, private readonly produtoService: ProdutosService) {}
+  constructor(
+    private readonly prismaService: PrismaService,
+    private readonly produtoService: ProdutosService,
+  ) {}
   async countAll() {
     return await this.prismaService.cliente.count({});
   }
@@ -18,10 +21,10 @@ export class OrcamentosService {
   async findAllWithPagination(page: number, perPage: number) {
     const skip = (page - 1) * perPage;
     const orcamentos = await this.prismaService.orcamento.findMany({
-    skip,
-    take: perPage,
-  });
-  return { orcamentos };
+      skip,
+      take: perPage,
+    });
+    return { orcamentos };
   }
 
   async create(createOrcamentoDto: CreateOrcamentoDto) {
@@ -47,30 +50,46 @@ export class OrcamentosService {
   }
 
   async update(id: number, updateOrcamentoDto: UpdateOrcamentoDto) {
-    const orcamento = await this.prismaService.orcamento.update({
-      where: { id },
-      data: updateOrcamentoDto,
-    });
-
-    if (orcamento) {
-      return orcamento;
+    const orcamentoExists = await this.findOne(id);
+    if (orcamentoExists) {
+      const clienteExists = await this.findCliente(
+        updateOrcamentoDto.idCliente,
+      );
+      if (clienteExists) {
+        if (orcamentoExists.status === 'Concluido') {
+          return await this.prismaService.orcamento.update({
+            where: { id },
+            data: {
+              status: updateOrcamentoDto.status,
+            },
+          });
+        }
+        return await this.prismaService.orcamento.update({
+          where: { id },
+          data: updateOrcamentoDto,
+        });
+      }
+      return { data: { message: 'Cliente não existe' } };
     }
-
-    return { data: { message: 'Ocorreu um erro ao atualizar o orcamento' } };
+    return { data: { message: 'Orçamento não existe' } };
   }
 
   async remove(id: number) {
-    const produtos = await this.prismaService.produto.findMany({
-      where: {
-        orcamentoId: id,
-      },
-    });
-    for (const produto of produtos) {
-      await this.produtoService.remove(produto.id)
+    const orcamentoExists = await this.findOne(id);
+    if (orcamentoExists) {
+      const produtos = await this.prismaService.produto.findMany({
+        where: {
+          orcamentoId: id,
+        },
+      });
+      for (const produto of produtos) {
+        await this.produtoService.remove(produto.id);
+      }
+      const removeOrcamento = await this.prismaService.orcamento.delete({
+        where: { id },
+      });
+      return { removeOrcamento };
     }
-    const removeOrcamento = await this.prismaService.orcamento.delete({
-      where: { id },
-    });
-    return { removeOrcamento };
+    return { data: { message: 'Orçamento não existe' } };
   }
 }
